Added the option facility.
As of now the code gives out options to the user. This will help reduce
errors in the spelling. But the next step will be to give only four not
ten options. And these four options will have to include the answer.

The options are displayed once at the beginning. And then the code
randomly chooses the which flags to display.